bible-app-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.5.0
pydantic==2.4.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
firebase-admin==6.2.0
redis==5.0.1
alembic==1.13.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.23.2
